Refactor: Consolidate sumcount aggregate resolver logic
Move aggregate_resolver_sumcount_only.cpp to aggregate_resolver_sumcount.cpp to reduce code duplication and improve maintainability.
