+++ be/src/exprs/CMakeLists.txt
@@ -26,7 +26,7 @@ set(EXPR_FILES
   agg/factory/aggregate_resolver_hypothesis_testing.cpp
   agg/factory/aggregate_resolver_minmaxany.cpp
   agg/factory/aggregate_resolver_others.cpp
-  agg/factory/aggregate_resolver_sumcount_only.cpp
+  agg/factory/aggregate_resolver_sumcount.cpp
   agg/factory/aggregate_resolver_distinct.cpp
   agg/factory/aggregate_resolver_stream.cpp
   agg/factory/aggregate_resolver_utility.cpp
